Move cache methods to cache class
